Conflicts:
	include/pdal/Dimension.hpp
	include/pdal/PointBuffer.hpp
	include/pdal/Schema.hpp
	include/pdal/drivers/oci/common.hpp
	include/pdal/drivers/qfit/Reader.hpp
	include/pdal/drivers/sbet/Common.hpp
	src/PointBuffer.cpp
	src/drivers/las/Reader.cpp
	src/drivers/qfit/Reader.cpp
	src/drivers/sqlite/Writer.cpp
	src/kernel/Delta.cpp

+++ src/PointBuffer.cpp
@@ -114,13 +114,13 @@ std::ostream& PointBuffer::toRST(std::ostream& os) const
 {
     const Dimension::IdList dims = m_context.dims();
 
-    uint32_t name_column(20);
-    uint32_t value_column(40);
+    size_t name_column(20);
+    size_t value_column(40);
     
     for (auto di = dims.begin(); di != dims.end(); ++di)
     {
         std::string name = Dimension::name(*di);
-        name_column = std::max(name_column, (uint32_t)name.size());
+        name_column = std::max(name_column, name.size());
     }
     
     std::ostringstream thdr;
